포신 각도 수정

+++ Team_1/Team_1/Obj.cpp
@@ -14,8 +14,8 @@ CObj::~CObj()
 
 void CObj::Update_Rect(void)
 {
-	m_tRC.left = (m_tInfo.fX - (m_tInfo.fCX * 0.5f));
-	m_tRC.top = (m_tInfo.fY - (m_tInfo.fCY * 0.5f));
-	m_tRC.right = (m_tInfo.fX + (m_tInfo.fCX * 0.5f));
-	m_tRC.bottom = (m_tInfo.fY + (m_tInfo.fCY * 0.5f));
+	m_tRC.left = LONG(m_tInfo.fX - (m_tInfo.fCX * 0.5f));
+	m_tRC.top = LONG(m_tInfo.fY - (m_tInfo.fCY * 0.5f));
+	m_tRC.right = LONG(m_tInfo.fX + (m_tInfo.fCX * 0.5f));
+	m_tRC.bottom = LONG(m_tInfo.fY + (m_tInfo.fCY * 0.5f));
 }
